Break out OrgLayout variation into separate component SimpleOrgLayout, rename OrgLayout
