+++ src/pages/o/[orgId].tsx
@@ -3,7 +3,7 @@ import { GetServerSideProps } from 'next';
 import { QueryClient, useQuery } from 'react-query';
 
 import getOrg from '../../fetching/getOrg';
-import OrgLayout from '../../components/layout/OrgLayout';
+import MainOrgLayout from '../../components/layout/MainOrgLayout';
 import { PageWithLayout } from '../../types';
 import { scaffold } from '../../utils/next';
 
@@ -48,9 +48,9 @@ const OrgPage : PageWithLayout<OrgPageProps> = (props) =>{
 
 OrgPage.getLayout = function getLayout(page, props) {
     return (
-        <OrgLayout mainPage={ true } orgId={ props.orgId as string }>
+        <MainOrgLayout orgId={ props.orgId as string }>
             { page }
-        </OrgLayout>
+        </MainOrgLayout>
     );
 };
