+++ src/pages/o/[orgId]/campaigns.tsx
@@ -6,7 +6,7 @@ import { QueryClient, useQuery } from 'react-query';
 
 import getCampaigns from '../../../fetching/getCampaigns';
 import getOrg from '../../../fetching/getOrg';
-import OrgLayout from '../../../components/layout/OrgLayout';
+import MainOrgLayout from '../../../components/layout/MainOrgLayout';
 import { PageWithLayout } from '../../../types';
 import { scaffold } from '../../../utils/next';
 
@@ -69,9 +69,9 @@ const OrgCampaignsPage : PageWithLayout<OrgCampaignsPageProps> = (props) => {
 
 OrgCampaignsPage.getLayout = function getLayout(page, props) {
     return (
-        <OrgLayout mainPage={ true } orgId={ props.orgId as string }>
+        <MainOrgLayout orgId={ props.orgId as string }>
             { page }
-        </OrgLayout>
+        </MainOrgLayout>
     );
 };
