+++ src/pages/o/[orgId]/surveys/[surId].tsx
@@ -4,9 +4,9 @@ import { QueryClient, useQuery } from 'react-query';
 
 import getOrg from '../../../../fetching/getOrg';
 import getSurvey from '../../../../fetching/getSurvey';
-import OrgLayout from '../../../../components/layout/OrgLayout';
 import { PageWithLayout } from '../../../../types';
 import { scaffold } from '../../../../utils/next';
+import SimpleOrgLayout from '../../../../components/layout/SimpleOrgLayout';
 
 export const getServerSideProps : GetServerSideProps = scaffold(async (context) => {
     const queryClient = new QueryClient();
@@ -55,9 +55,9 @@ const OrgSurveyPage : PageWithLayout<OrgSurveyPageProps> = (props) => {
 
 OrgSurveyPage.getLayout = function getLayout(page, props) {
     return (
-        <OrgLayout mainPage={ false } orgId={ props.orgId as string }>
+        <SimpleOrgLayout orgId={ props.orgId as string }>
             { page }
-        </OrgLayout>
+        </SimpleOrgLayout>
     );
 };
